Consider updating this documentation (FTX)
I believe FTX is not part of the Data Availability Committee anymore. Is that right? Moreover, I believe that linking Arbitrum to FTX isn't beneficial to the project. Thus, I propose a slight change to the documentation to preserve and protect Arbitrum's image and reputation, as well as reflect the Data Committee more accurately in sight of the recent events.